:white_check_mark: refactor : 피드백 반영

+++ src/main/java/kr/hhplus/be/server/coupon/domain/service/response/CouponsResponse.java
@@ -7,6 +7,6 @@ import java.time.LocalDateTime;
 public record CouponsResponse(
         long couponId,
         String couponName,
-        long couponQuantity
+        int couponQuantity
 ) {
 }
